Move quotations and contexts to a slightly better home.
The correct solution is probably to permanently eliminate QuoteLink
and LocalQuoteLink.  In retrospect, they were calamitously bad design
ideas.
