+++ opencog/atoms/pattern/PatternTerm.h
@@ -31,7 +31,7 @@
 
 #include <opencog/atoms/base/Handle.h>
 #include <opencog/atoms/base/Link.h>
-#include <opencog/atoms/base/Quotation.h>
+#include <opencog/atoms/core/Quotation.h>
 
 namespace opencog {
