+++ opencog/atomutils/FindUtils.h
@@ -37,7 +37,7 @@
 #include <opencog/atoms/base/Handle.h>
 #include <opencog/atoms/base/Link.h>
 #include <opencog/atoms/base/types.h>
-#include <opencog/atoms/base/Quotation.h>
+#include <opencog/atoms/core/Quotation.h>
 #include <opencog/atoms/core/ScopeLink.h>
 
 namespace opencog {
