+++ opencog/rule-engine/Rule.cc
@@ -27,13 +27,13 @@
 #include <boost/uuid/uuid_io.hpp>
 #include <boost/uuid/uuid_generators.hpp>
 
-#include <opencog/atomspace/AtomSpace.h>
-#include <opencog/atoms/base/Node.h>
 #include <opencog/atoms/base/Link.h>
-#include <opencog/atoms/base/Quotation.h>
+#include <opencog/atoms/base/Node.h>
 #include <opencog/atoms/core/DefineLink.h>
+#include <opencog/atoms/core/Quotation.h>
 #include <opencog/atoms/pattern/BindLink.h>
 
+#include <opencog/atomspace/AtomSpace.h>
 #include <opencog/atomutils/TypeUtils.h>
 #include <opencog/atomutils/Unify.h>
