+++ tests/atomutils/UnifyUTest.cxxtest
@@ -23,8 +23,8 @@
 
 #include <opencog/util/Logger.h>
 
+#include <opencog/atoms/core/Context.h>
 #include <opencog/atomutils/Unify.h>
-#include <opencog/atoms/base/Context.h>
 #include <opencog/atomspace/AtomSpace.h>
 #include <opencog/guile/SchemeEval.h>
